omero3 compatibility
git-svn-id: file:///home/svn/shoola-omero/trunk@3243 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/hiviewer/CategoryGroupLoader.java
@@ -63,7 +63,7 @@ public class CategoryGroupLoader
 {
 
     /** The id of the root Category Group. */
-    private int         cgID;
+    private long         cgID;
     
     /** Handle to the async call so that we can cancel it. */
     private CallHandle  handle;
@@ -76,7 +76,7 @@ public class CategoryGroupLoader
      *               Mustn't be <code>null</code>.
      * @param cgID  The id of the root Category Group.
      */
-    public CategoryGroupLoader(HiViewer viewer, int cgID)
+    public CategoryGroupLoader(HiViewer viewer, long cgID)
     {
         super(viewer);
         this.cgID = cgID;
